Rename viewmodel.ts -> project-data-viewmodel.ts
