+++ monosketch-svelte/src/lib/ui/modal/recent-project/project-list/ProjectList.svelte
@@ -3,7 +3,7 @@ import ProjectRow from './ProjectRow.svelte';
 import { ProjectAction, type ProjectItem } from '../model';
 import { onDestroy, onMount } from 'svelte';
 import { LifecycleOwner } from '$libs/flow';
-import { projectDataViewModel } from '../viewmodel';
+import { projectDataViewModel } from '$ui/modal/recent-project/project-data-viewmodel';
 import { BrowserManager } from "$mono/window/browser-manager";
 
 export let dismiss: () => void;
