+++ monosketch-svelte/src/lib/ui/nav/NavBar.svelte
@@ -7,6 +7,8 @@ import Brand from './brand/Brand.svelte';
 import MouseActionGroup from './mouseaction/MouseActionGroup.svelte';
 import ProjectManagerButton from './project/ProjectManagerButton.svelte';
 import WorkingProject from './project/WorkingProject.svelte';
+
+
 </script>
 
 <div class="container">
