+++ monosketch-svelte/src/lib/ui/nav/project/WorkingProjectView.svelte
@@ -3,7 +3,7 @@
     import { TargetBounds } from '$ui/modal/model';
     import CurrentProjectDropDown from "$ui/modal/menu/current-project/CurrentProjectDropDown.svelte";
     import { onMount } from "svelte";
-    import { projectDataViewModel } from "$ui/modal/recent-project/viewmodel";
+    import { projectDataViewModel } from '$ui/modal/recent-project/project-data-viewmodel';
     import { LifecycleOwner } from "$libs/flow";
     import RenameProjectModal from "$ui/modal/rename-project/RenameProjectModal.svelte";
